added functions: register, authorithation, authentification from email

+++ src/main/java/com/example/buysell/models/Product.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -29,7 +28,7 @@ public class Product {
     private String description;
 
     @Column(name = "price")
-    private int price;
+    private Double price;
 
     @Column(name = "city")
     private String city;
